refactor(products): optimize InfiniteScrollProducts with useInfiniteScrollProducts hook and ProductListWrapper component
